Backend/requirements.txt
fastapi==0.99.1
uvicorn[standard]==0.22.0
python-dotenv==1.0.0
supabase==1.0.0
httpx==0.24.1
